Refactor money and stats handling to use long long for improved accuracy; update related functions and file operations

+++ blackjackfolder/printingsystem.h
@@ -11,7 +11,7 @@ extern const char *colors[];
 
 int startupscreen();
 void leaderboard();
-void stats(int *overallplayed, int *totalwins, int *totallost, int *totalties, int *moneygained, int *moneylost);
+void stats(long long *overallplayed, long long *totalwins, long long *totallost, long long *totalties, long long *moneygained, long long *moneylost);
 void print_deck(int deck[][14]);
 void draw_cards(int how_many, int hand[][20]);
 void carddraw(int deck[][14], int hand[][20], int *limit, int *sum, int *played_cards, int *softflag);
